% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_packages.R
\name{required_packages}
\alias{required_packages}
\title{Required Packages}
\usage{
required_packages()
}
\value{
A character vector of required packages.
}
\description{
A function to list required packages.
}
