% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taso_colours.R
\name{taso_colours}
\alias{taso_colours}
\title{Get TASO colours without specifying number}
\usage{
taso_colours(levels)
}
\arguments{
\item{levels}{A character vector of levels for which colours are needed.}
}
\value{
A named character vector of colours.
}
\description{
This function returns the custom TASO colours mapped to the levels of a factor.
}
\examples{
levels <- c("A", "B", "C", "D")
taso_colours(levels)
}
