% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taso_colours.R
\docType{data}
\name{tasocolours}
\alias{tasocolours}
\title{TASO Custom Color Palette}
\format{
An object of class \code{character} of length 6.
}
\usage{
tasocolours
}
\description{
This vector defines the custom color palette for TASO.
}
\keyword{datasets}
